refactor: 重构为纯x86

+++ src/mythroad/src/h/mr_limits.h
@@ -115,7 +115,7 @@ typedef MRP_UACNUMBER l_uacNumber;
 ** type for virtual-machine instructions
 ** must be an unsigned with (at least) 4 bytes (see details in lopcodes.h)
 */
-typedef unsigned long Instruction;
+typedef unsigned int Instruction;
 
 
 /* maximum depth for calls (unsigned short) */
